+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Rename `linkSourceGraphToTargetGraph(...)` to `bulkLinkFromSourceGraphToTargetGraph(...)` in `AbstractSessionContextImpl` to match conventions present in domgen generated code.
 * Add initial support for "bulk" loading of type graphs. While "bulk" loading is a bit of a misnomer given that type graphs have at most 1 subscription, the term is synonymous with loading using SQL queries in domgen generated code.
 * Remove return value from bulk loading methods as implementations never return false, they generate an exception or perform the bulk load.
+* Fix a bug in `AbstractSessionContextImpl.bulkLinkFromSourceGraphToTargetGraph(...)` where the ids used in linking from the source graph to the target graph were incorrectly inverted.
 
 ### [v6.117](https://github.com/replicant4j/replicant/tree/v6.117) (2022-03-29) · [Full Changelog](https://github.com/spritz/spritz/compare/v6.116...v6.117)